Rename `domain` test package to `examples`
