--- src/test/java/com/fluentinterface/domain/examples/Customer.java
@@ -1,1 +1,0 @@
-package com.fluentinterface.domain.examples;public class Customer extends HumanCommon {	private CustomerType type;	public CustomerType getType() {		return type;	}	public void setType(CustomerType type) {		this.type = type;	}}
