--- src/test/java/com/fluentinterface/domain/examples/CustomerBuilder.java
@@ -1,1 +1,0 @@
-package com.fluentinterface.domain.examples;import com.fluentinterface.builder.Builder;public interface CustomerBuilder          extends HumanCommonBuilder<CustomerBuilder>,                 Builder<Customer> {	CustomerBuilder withType(final CustomerType type);		/** Setting unknown properties will fail and not ascend to the Object */	CustomerBuilder withAnUnknownProperty(final String value);}
