--- src/test/java/com/fluentinterface/domain/examples/CustomerType.java
@@ -1,1 +1,0 @@
-package com.fluentinterface.domain.examples;public enum CustomerType {	BASIC, SILVER, GOLD}
