--- src/test/java/com/fluentinterface/domain/examples/DbCommon.java
@@ -1,1 +1,0 @@
-package com.fluentinterface.domain.examples;import java.util.Calendar;public abstract class DbCommon {	private String id;	private int version;	private String description;	private Calendar created;	private String createdBy;	private Calendar updated;	private String updatedBy;	public String getId() {		return id;	}	public void setId(String id) {		this.id = id;	}	public int getVersion() {		return version;	}	public void setVersion(int version) {		this.version = version;	}	public String getDescription() {		return description;	}	public void setDescription(String description) {		this.description = description;	}	public Calendar getCreated() {		return created;	}	public void setCreated(Calendar created) {		this.created = created;	}	public String getCreatedBy() {		return createdBy;	}	public void setCreatedBy(String createdBy) {		this.createdBy = createdBy;	}	public Calendar getUpdated() {		return updated;	}	public void setUpdated(Calendar updated) {		this.updated = updated;	}	public String getUpdatedBy() {		return updatedBy;	}	public void setUpdatedBy(String updatedBy) {		this.updatedBy = updatedBy;	}}
