--- src/test/java/com/fluentinterface/domain/examples/EmployeeBuilder.java
@@ -1,1 +1,0 @@
-package com.fluentinterface.domain.examples;import com.fluentinterface.builder.Builder;public interface EmployeeBuilder          extends HumanCommonBuilder<EmployeeBuilder>,                 Builder<Employee> {	EmployeeBuilder withDepartment(final String department);		EmployeeBuilder withSalary(final double salary);}
