--- src/test/java/com/fluentinterface/domain/examples/HumanCommon.java
@@ -1,1 +1,0 @@
-package com.fluentinterface.domain.examples;public abstract class HumanCommon extends DbCommon {	private String firstName;	private String lastName;	private String address;	public String getAddress() {		return address;	}	public void setAddress(String address) {		this.address = address;	}	public String getFirstName() {		return firstName;	}	public void setFirstName(String firstName) {		this.firstName = firstName;	}	public String getLastName() {		return lastName;	}	public void setLastName(String lastName) {		this.lastName = lastName;	}}
