--- src/test/java/com/fluentinterface/domain/examples/HumanCommonBuilder.java
@@ -1,1 +1,0 @@
-package com.fluentinterface.domain.examples;public interface HumanCommonBuilder<T extends HumanCommonBuilder<T>>          extends DbCommonBuilder<T> {		T withFirstName(final String firstName);	T withLastName(final String lastName);	T withAddress(final String address);}
