+++ src/test/java/com/fluentinterface/examples/TypedStringIDBuilder.java
@@ -1,4 +1,4 @@
-package com.fluentinterface.domain;
+package com.fluentinterface.examples;
 
 import com.fluentinterface.builder.Builder;
