+++ src/test/java/com/fluentinterface/examples/contrib/Customer.java
@@ -1,0 +1,1 @@
+package com.fluentinterface.examples.contrib;public class Customer extends HumanCommon {	private CustomerType type;	public CustomerType getType() {		return type;	}	public void setType(CustomerType type) {		this.type = type;	}}
