+++ src/test/java/com/fluentinterface/examples/contrib/CustomerBuilder.java
@@ -1,0 +1,1 @@
+package com.fluentinterface.examples.contrib;import com.fluentinterface.builder.Builder;public interface CustomerBuilder          extends HumanCommonBuilder<CustomerBuilder>,                 Builder<Customer> {	CustomerBuilder withType(final CustomerType type);		/** Setting unknown properties will fail and not ascend to the Object */	CustomerBuilder withAnUnknownProperty(final String value);}
