+++ src/test/java/com/fluentinterface/examples/contrib/CustomerType.java
@@ -1,0 +1,1 @@
+package com.fluentinterface.examples.contrib;public enum CustomerType {	BASIC, SILVER, GOLD}
