+++ src/test/java/com/fluentinterface/examples/contrib/Employee.java
@@ -1,0 +1,1 @@
+package com.fluentinterface.examples.contrib;public class Employee extends HumanCommon {	private String department;	private double salary;	public String getDepartment() {		return department;	}	public void setDepartment(String department) {		this.department = department;	}	public double getSalary() {		return salary;	}	public void setSalary(double salary) {		this.salary = salary;	}}
