+++ src/test/java/com/fluentinterface/examples/contrib/HumanCommon.java
@@ -1,0 +1,1 @@
+package com.fluentinterface.examples.contrib;public abstract class HumanCommon extends DbCommon {	private String firstName;	private String lastName;	private String address;	public String getAddress() {		return address;	}	public void setAddress(String address) {		this.address = address;	}	public String getFirstName() {		return firstName;	}	public void setFirstName(String firstName) {		this.firstName = firstName;	}	public String getLastName() {		return lastName;	}	public void setLastName(String lastName) {		this.lastName = lastName;	}}
