+++ src/test/java/com/fluentinterface/examples/contrib/HumanCommonBuilder.java
@@ -1,0 +1,1 @@
+package com.fluentinterface.examples.contrib;public interface HumanCommonBuilder<T extends HumanCommonBuilder<T>>          extends DbCommonBuilder<T> {		T withFirstName(final String firstName);	T withLastName(final String lastName);	T withAddress(final String address);}
